requirements.txt
flask>=3.0.0
flask-session>=0.8.0
groq>=0.11.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-flask>=1.3.0
